feat(tabs): convert to Typescript
